+++ RELEASES.md
@@ -30,6 +30,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * prov-template-compiler: @semantic.type annotation, allowing semantic type of template to be found easily
 * service-template-core/archetype: backward traversal version returning semantic type of activity, and update of viz 
 * prov-template-archetype: T19 update navigation display to visualise activity type icons
+* prov-explanations: extract json property
 
 `2.2.3`
 * prov-template: update to instantiation